rename example to quantization-qdqbert; rename example scripts from qat to quant
